requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.8.1
altair==4.2.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asteroid-filterbanks==0.4.0
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
audioread==3.0.0
autopage==0.5.1
backcall==0.2.0
backports.cached-property==1.0.2
beautifulsoup4==4.11.1
black==22.12.0
bleach==5.0.1
blinker==1.5
cachetools==5.2.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cliff==4.0.0
cmaes==0.9.0
cmd2==2.4.2
colorama==0.4.6
colorlog==6.7.0
commonmark==0.9.1
config==0.5.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.3
decorator==4.4.2
defusedxml==0.7.1
docopt==0.6.2
einops==0.3.2
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
ffmpeg-python==0.2.0
filelock==3.8.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.2
gdown==4.5.3
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.14.1
google-auth-oauthlib==0.4.6
greenlet==2.0.1
grpcio==1.50.0
hmmlearn==0.2.8
huggingface-hub==0.10.1
HyperPyYAML==1.0.1
idna==3.4
imageio==2.22.4
imageio-ffmpeg==0.4.7
importlib-metadata==4.13.0
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
julius==0.2.7
jupyter-server==1.23.2
jupyter_client==7.4.6
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
librosa==0.9.2
llvmlite==0.39.1
Mako==1.2.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
methodtools==0.4.5
mistune==2.0.4
more-itertools==9.0.0
moviepy==1.0.3
mpmath==1.2.1
multidict==6.0.2
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
nltk==3.7
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.2.3
openai==0.25.0
opencv-python==4.6.0.66
openpyxl==3.0.10
optuna==3.0.3
packaging==21.3
pandas==1.5.1
pandas-stubs==1.5.1.221024
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.3
pbr==5.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
pooch==1.6.0
prettytable==3.5.0
primePy==1.3
proglog==0.1.10
prometheus-client==0.15.0
prompt-toolkit==3.0.32
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.audio==2.1.1
pyannote.core==4.5
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyperclip==1.8.2
pyrouge==0.1.3
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-speech-features==0.6
pytorch-lightning==1.6.5
pytorch-metric-learning==1.6.3
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rich==12.6.0
rouge==1.0.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scenedetect==0.6.0.3
scikit-learn==1.1.3
scipy==1.8.1
semver==2.13.0
Send2Trash==1.8.0
sentencepiece==0.1.97
shellingham==1.5.0
simplejson==3.18.0
singledispatchmethod==1.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
speechbrain==0.5.13
SQLAlchemy==1.4.44
stack-data==0.6.1
stevedore==4.1.1
streamlit==1.16.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.0
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.2
torchaudio==0.13.0
torchmetrics==0.10.2
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.24.0
typer==0.7.0
types-pytz==2022.6.0.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uisrnn==0.1.0
urllib3==1.26.12
validators==0.20.0
watchdog==2.2.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
whisper @ git+https://github.com/openai/whisper.git@9f70a352f9f8630ab3aa0d06af5cb9532bd8c21d
widgetsnbextension==4.0.3
wirerope==0.4.6
yarl==1.8.1
youtube-dl==2021.12.17
zipp==3.10.0
